query_data/requirements.txt
mysql.connector
selenium
